data_processing/requirements.txt
pinecone-client
pandas
python-dotenv
openai
